+++ packages/mobile/src/alpha/combobox/Combobox.tsx
@@ -67,7 +67,7 @@ export type ComboboxControlProps<
   Type extends SelectType = 'single',
   SelectOptionValue extends string = string,
 > = SelectControlProps<Type, SelectOptionValue> &
-  Pick<ComboboxBaseProps<Type, SelectOptionValue>, 'hideSearchInput'> & {
+  Pick<ComboboxBaseProps<Type, SelectOptionValue>, 'hideSearchInput' | 'font'> & {
     /** Search text value */
     searchText: string;
     /** Search text change handler */
@@ -196,6 +196,7 @@ const ComboboxBase = memo(
         ComboboxControlComponent = DefaultComboboxControl,
         SelectDropdownComponent = DefaultSelectDropdown,
         hideSearchInput,
+        font,
         ...props
       } = mergedProps;
       const [searchTextInternal, setSearchTextInternal] = useState(defaultSearchText);
@@ -264,11 +265,12 @@ const ComboboxBase = memo(
               ComboboxControlComponent={ComboboxControlComponent}
               SelectControlComponent={SelectControlComponent}
               controlRef={controlRef}
+              font={font}
               searchInputRef={searchInputRef}
             />
           );
         },
-        [ComboboxControlComponent, SelectControlComponent, searchInputRef],
+        [ComboboxControlComponent, SelectControlComponent, font, searchInputRef],
       );
 
       const ComboboxDropdown = useCallback(
@@ -309,6 +311,7 @@ const ComboboxBase = memo(
                   startNode={startNode}
                   variant={variant}
                   {...props}
+                  font={font}
                   label={null}
                   styles={undefined}
                 />
@@ -324,6 +327,7 @@ const ComboboxBase = memo(
           align,
           closeButtonLabel,
           endNode,
+          font,
           handleTrayVisibilityChange,
           label,
           placeholder,