+++ packages/mobile/src/controls/Control.tsx
@@ -35,6 +35,8 @@ export type ControlIconProps = SharedProps & {
   borderRadius?: ThemeVars.BorderRadius;
   borderWidth?: ThemeVars.BorderWidth;
   elevation?: ElevationLevels;
+  controlSize?: number;
+  dotSize?: number;
   animatedScaleValue: Animated.Value;
   animatedOpacityValue: Animated.Value;
   accessible?: boolean;
@@ -72,6 +74,15 @@ export type ControlBaseProps<ControlValue extends string> = Omit<
     controlColor?: ThemeVars.Color;
     /** Sets the elevation/drop shadow of the control. */
     elevation?: ElevationLevels;
+    /**
+     * Sets the control size in pixels.
+     */
+    controlSize?: number;
+    /**
+     * Sets the inner dot size in pixels.
+     * @default 2/3 of controlSize
+     */
+    dotSize?: number;
     style?: ViewStyle;
   };
 
@@ -116,6 +127,8 @@ const ControlWithRef = forwardRef(function ControlWithRef<ControlValue extends s
     borderColor,
     borderRadius,
     borderWidth,
+    controlSize,
+    dotSize,
     ...props
   } = mergedProps;
   const theme = useTheme();
@@ -233,7 +246,9 @@ const ControlWithRef = forwardRef(function ControlWithRef<ControlValue extends s
           borderWidth={borderWidth}
           checked={checked}
           controlColor={controlColor}
+          controlSize={controlSize}
           disabled={pressDisabled}
+          dotSize={dotSize}
           elevation={elevation}
           indeterminate={indeterminate}
           pressed={pressed}
@@ -267,7 +282,9 @@ const ControlWithRef = forwardRef(function ControlWithRef<ControlValue extends s
       borderWidth,
       checked,
       controlColor,
+      controlSize,
       disabled,
+      dotSize,
       elevation,
       getLabelStyle,
       iconWrapperStyles,