+++ packages/web/src/alpha/select/Select.tsx
@@ -107,6 +107,7 @@ const SelectBase = memo(
         media,
         end,
         align,
+        font,
         bordered = true,
         SelectOptionComponent = DefaultSelectOption,
         SelectAllOptionComponent = DefaultSelectAllOption,
@@ -276,6 +277,7 @@ const SelectBase = memo(
             compact={compact}
             disabled={disabled}
             endNode={endNode}
+            font={font}
             helperText={helperText}
             hiddenSelectedOptionsLabel={hiddenSelectedOptionsLabel}
             label={label}