+++ packages/web/src/controls/SearchInput.tsx
@@ -9,7 +9,7 @@ import { Box } from '../layout/Box';
 
 import { InputIcon } from './InputIcon';
 import { InputIconButton } from './InputIconButton';
-import { TextInput, type TextInputBaseProps } from './TextInput';
+import { TextInput, type TextInputBaseProps, type TextInputProps } from './TextInput';
 
 export const scales = {
   regular: 56,
@@ -24,20 +24,19 @@ const compactCss = css`
   height: ${scales.compact}px;
 `;
 
-type HTMLElementProps = React.InputHTMLAttributes<HTMLInputElement> &
-  Required<Pick<HTMLInputElement, 'value'>>;
-
 export type SearchInputBaseProps = Pick<
   TextInputBaseProps,
   | 'accessibilityHint'
   | 'accessibilityLabel'
   | 'accessibilityLabelledBy'
   | 'bordered'
+  | 'borderRadius'
   | 'compact'
   | 'disabled'
   | 'enableColorSurge'
   | 'focusedBorderWidth'
   | 'helperTextErrorIconAccessibilityLabel'
+  | 'font'
   | 'placeholder'
   | 'testID'
   | 'testIDMap'
@@ -64,12 +63,10 @@ export type SearchInputBaseProps = Pick<
   startIcon?: Extract<IconName, 'search' | 'backArrow'>;
   /**
    * hide the end icon
-   * @default undefined
    */
   hideEndIcon?: boolean;
   /**
    * Set the end node
-   * @default undefined
    */
   end?: React.ReactNode;
   /**
@@ -83,7 +80,7 @@ export type SearchInputBaseProps = Pick<
 };
 
 export type SearchInputProps = SearchInputBaseProps &
-  HTMLElementProps & {
+  TextInputProps & {
     onClear?: React.MouseEventHandler;
     onChangeText: (text: string) => void;
     /**
@@ -112,6 +109,7 @@ export const SearchInput = memo(
       end,
       startIconAccessibilityLabel = 'Back',
       clearIconAccessibilityLabel = 'Clear search query',
+      borderRadius = 1000,
       ...props
     } = mergedProps;
     const internalRef = useRef<HTMLInputElement>(null);
@@ -152,7 +150,7 @@ export const SearchInput = memo(
     return (
       <TextInput
         ref={refs}
-        borderRadius={1000}
+        borderRadius={borderRadius}
         className={cx(baseCss, compact && compactCss)}
         end={
           end ??