+++ packages/common/CHANGELOG.md
@@ -8,6 +8,18 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 8.59.0 ((3/27/2026, 05:43 AM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.58.0 ((3/25/2026, 11:42 AM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.57.1 ((3/24/2026, 01:14 PM PST))
+
+This is an artificial version bump with no new change.
+
 ## 8.57.0 ((3/24/2026, 12:46 PM PST))
 
 This is an artificial version bump with no new change.